+++ packages/admin-panel/src/components/LayoutAdministrators/SearchBar.jsx
