--- packages/admin-panel/src/components/ModalAddAdmin/ModalCreateAdminContext.jsx
@@ -1,66 +1,0 @@
-import React from "react";
-import useLocalStorage from "./useLocalStorage";
-const TodoContext = React.createContext();
-
-const TodoProvider = (props) => {
-  const {
-    item: todos,
-    saveItem: saveTodos,
-    loading,
-    error,
-  } = useLocalStorage("TODOS_V1", []);
-  const [openModal, setOpenModal] = React.useState(false);
-  const [searchValue, setSearchValue] = React.useState("");
-  const completedTodos = todos.filter((todo) => !!todo.completed).length;
-  const totalTodos = todos.length;
-  let searchedTodos = [];
-  if (!searchValue.length >= 1) {
-    searchedTodos = todos;
-  } else {
-    searchedTodos = todos.filter((todo) => {
-      const todoText = todo.title.toLowerCase();
-      const searchText = searchValue.toLowerCase();
-      return todoText.includes(searchText);
-    });
-  }
-
-  const completeTodo = (props) => {
-    const todoIndex = todos.findIndex((todo) => todo.title === props);
-    const newTodos = [...todos];
-    newTodos[todoIndex].completed = !newTodos[todoIndex].completed;
-    saveTodos(newTodos);
-  };
-
-  const addTodo = (props) => {
-    console.log(props);
-  };
-
-  const deleteTodo = (props) => {
-    const todoIndex = todos.findIndex((todo) => todo.title === props);
-    const newTodos = [...todos];
-    newTodos.splice(todoIndex, 1);
-    saveTodos(newTodos);
-  };
-  return (
-    <TodoContext.Provider
-      value={{
-        completedTodos,
-        totalTodos,
-        searchValue,
-        setSearchValue,
-        searchedTodos,
-        completeTodo,
-        deleteTodo,
-        loading,
-        error,
-        openModal,
-        setOpenModal,
-        addTodo,
-      }}
-    >
-      {props.children}
-    </TodoContext.Provider>
-  );
-};
-
-export { TodoContext, TodoProvider };
