--- packages/admin-panel/src/components/Paginate.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const Paginate = (props) => {
-  return <div>Paginate</div>;
-};
